Only try to serialize PhoneNumber instances
AbstractBillingAddress has no phone_number field. But presumably because it is common to add it, we have code dealing with it's existence. The existing code assumed that a PhoneNumberField model field would be added, which might not always be the case.
I stumbled upon this in a project of mine where phone_number was a CharField (for reasons...). That would break with an AttributeError.
The fix is easy enough, we just check if we're indeed dealing with a PhoneNumber instance before calling `as_international` on it.

Note that this fix does not need to be applied to shipping addresses, because there we actually define a PhoneNumberField, so our assumptions hold.